ad-1 book returns, bug fixing related to reservations, handling exceptions, quering for priority reservations.

+++ src/main/java/com/university/bibliotheca/service/ReservationService.java
@@ -2,20 +2,25 @@ package com.university.bibliotheca.service;
 
 import com.university.bibliotheca.adapter.mongo.MongoReservationQueueAdapter;
 import com.university.bibliotheca.adapter.mongo.exception.AvailableBookNotFoundException;
-import com.university.bibliotheca.adapter.mongo.exception.ReservationNotFoundException;
+import com.university.bibliotheca.adapter.mongo.exception.ReservationQueueNotFoundException;
 import com.university.bibliotheca.domain.model.Book;
 import com.university.bibliotheca.domain.model.BorrowResult;
 import com.university.bibliotheca.domain.model.Reservation;
 import com.university.bibliotheca.domain.model.ReservationQueue;
 import com.university.bibliotheca.domain.model.ReturnResult;
 import com.university.bibliotheca.domain.model.User;
+import com.university.bibliotheca.service.exception.ReservationNotFoundException;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 
 import java.time.Instant;
 import java.util.Date;
 import java.util.List;
+import java.util.stream.Collectors;
+
+import static java.time.temporal.ChronoUnit.DAYS;
 
 @Log4j2
 @Service
@@ -25,6 +30,9 @@ public class ReservationService {
     private UserService userService;
 
     private MongoReservationQueueAdapter mongoReservationQueueAdapter;
+
+    @Value("${borrow_days}")
+    private int BORROW_DAYS;
 
     @Autowired
     public ReservationService(BookService bookService, UserService userService, MongoReservationQueueAdapter mongoReservationQueueAdapter) {
@@ -39,10 +47,15 @@ public class ReservationService {
 //    DONE.
     public BorrowResult borrowBook(String userId, String bookName, Date borrowEnd) {
         try {
+            if (!isAlreadyBorrowed(userId, bookName)) {
             Book bookToBorrow = bookService.findAvailableBookByName(bookName);
             bookService.changeBorrowStatus(bookToBorrow.getId(), true, userId, borrowEnd);
+
             userService.addBorrowToUser(userId, bookToBorrow.getId());
             return BorrowResult.BORROWED;
+        } else {
+                return BorrowResult.ALREADY_BORROWED;
+            }
         } catch (AvailableBookNotFoundException e) {
             log.info("[ReservationService] Didn't borrow book for user: " + userId + ", making reservation for book: " + bookName);
             if (isAlreadyReserved(userId, bookName)) {
@@ -56,21 +69,19 @@ public class ReservationService {
 
     public ReturnResult returnBook(String userId, String bookId) {
         Book returnedBook = bookService.findBook(bookId);
-        bookService.changeBorrowStatus(returnedBook.getId(), false, null, null);
-        userService.removeBorrowFromUser(userId, bookId);
+        if (returnedBook.getBorrower() != null && returnedBook.getBorrower().equals(userId)) {
+            bookService.changeBorrowStatus(returnedBook.getId(), false, null, null);
+            userService.removeBorrowFromUser(userId, bookId);
 
-        try {
-            findPriorityReservation();
-            //Zwrotka Returned and someone Borrowed from Reserved List
-        } catch (ReservationNotFoundException ignored){
-            //Zwrotka Returned and Not Borrowed from Reserved List
+            try {
+                borrowReservedBook(returnedBook);
+                return ReturnResult.RETURNED_AND_RESERVED;
+            } catch (ReservationQueueNotFoundException | ReservationNotFoundException e) {
+                return ReturnResult.RETURNED;
+            }
+        } else {
+            return ReturnResult.NOT_OWNED;
         }
-
-        return null;
-    }
-
-    public void findPriorityReservation(){
-        mongoReservationQueueAdapter.findPriorityReservation("test-book-name");
     }
 
     private void reserveBook(String userId, String bookName) {
@@ -86,9 +97,61 @@ public class ReservationService {
         userService.addReservationToUser(retrievedUser, bookName);
     }
 
+    private void borrowReservedBook(Book returnedBook) {
+        ReservationQueue reservationQueue = mongoReservationQueueAdapter.findQueue(returnedBook.getName());
+
+        assert reservationQueue != null;
+        if (reservationQueue.getUserReservations() != null && !reservationQueue.getUserReservations().isEmpty()) {
+            Reservation priorityReservation = findPriorityReservation(reservationQueue.getUserReservations());
+
+            User priorityUser = userService.findUser(priorityReservation.getUserId());
+
+            removeReservationFromQueue(reservationQueue, priorityReservation);
+            userService.removeReservationFromUser(priorityUser, returnedBook.getName());
+
+            borrowBook(priorityUser.getId(), returnedBook.getName(), Date.from(Instant.now().plus(BORROW_DAYS, DAYS)));
+        } else {
+            throw (new ReservationQueueNotFoundException(returnedBook.getName()));
+        }
+    }
+
+    private void removeReservationFromQueue(ReservationQueue reservationQueue, Reservation reservation) {
+        ReservationQueue updatedReservationQueue = removeReservation(reservationQueue, reservation);
+        mongoReservationQueueAdapter.saveReservationQueue(updatedReservationQueue);
+    }
+
+
+    private Reservation findPriorityReservation(List<Reservation> reservations) {
+//        return mongoReservationQueueAdapter.findPriorityReservation(List<Reservation> reservations);
+        //Returns Oldest Reservation
+        reservations.sort((o1, o2) -> {
+            if (o1.getReservationDate() == null || o2.getReservationDate() == null) {
+                return 0;
+            } else
+                return o1.getReservationDate().compareTo(o2.getReservationDate());
+        });
+
+        //Returns Higest Priority Occupation
+        reservations.sort((o1, o2) -> {
+            if (o1.getOccupation() == null || o2.getOccupation() == null) {
+                return 0;
+            } else
+                return o2.getOccupation().getNumVal().compareTo(o1.getOccupation().getNumVal());
+        });
+
+        return reservations.get(0);
+    }
+
     private boolean isAlreadyReserved(String userId, String bookName) {
         Reservation retrievedReservation = mongoReservationQueueAdapter.findReservation(bookName, userId);
         return (retrievedReservation != null);
+    }
+
+    private boolean isAlreadyBorrowed(String userId, String bookName) {
+        return userService.findUser(userId).getBorrowedBookIds().stream().map(bookId ->
+                bookService.findBook(bookId).getName()
+        ).collect(Collectors.toList())
+                .contains(bookName);
     }
 
     private void saveReservationQueue(ReservationQueue reservationQueue) {
@@ -109,5 +172,13 @@ public class ReservationService {
 
         return new ReservationQueue(reservationQueue.getName(), reservations);
     }
+
+    private ReservationQueue removeReservation(ReservationQueue reservationQueue, Reservation reservation) {
+        List<Reservation> reservations = reservationQueue.getUserReservations();
+        reservations.remove(reservation);
+
+        return new ReservationQueue(reservationQueue.getName(), reservations);
+    }
+
 }
